Rename Parser.py to not conflict with standard library package - caused me issues
