+++ .gitignore
@@ -1,0 +1,4 @@
+Lib/
+Scripts/
+__pycache__/
+pyvenv.cfg
