+++ main.py
@@ -1,13 +1,13 @@
 import random
 import wiki_downloader
-import parser
+import sentence_parser
 
 
 def get_random_article(article_list):
     article_title = random.choice(article_list)
     article = wiki_downloader.get_wikipedia_full_text(article_title)
     article_title.replace("_"," ")
-    sentences = parser.split_and_filter_text(article, article_title)
+    sentences = sentence_parser.split_and_filter_text(article, article_title)
     sentences = [s for s in sentences if article_title.lower() not in s.lower()]
     return article_title, sentences
 
@@ -39,6 +39,6 @@ def article_guessing_game(article_list):
     print("Out of guesses. The correct answer was:", correct_article)
 
 # Example usage:
-with open("articles.txt", "r") as f:
+with open("articles.txt", "r", errors="ignore") as f:
     article_names = [line.strip() for line in f]
 article_guessing_game(article_names)
